requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
python-dotenv
openrouter
tiktoken
pypdf
fastapi
uvicorn
motor
pydantic
crawl4ai
tavily-python
python-multipart
celery
redis
celery
redis
